requirements.txt
pymodbus>=3.6,<4
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic>=1.13
sqlalchemy>=2.0
